+++ packages/client/components/DialogContainer.ts
@@ -8,7 +8,7 @@ const DialogContainer = styled('div')({
   borderRadius: Radius.DIALOG,
   boxShadow: modalShadow,
   flexDirection: 'column',
-  // overflow: 'auto', removed because react-beautiful-dnd only supports 1 scrolling parent
+  // overflow: 'auto', removed because @hello-pangea/dnd only supports 1 scrolling parent
   margin: '0 auto',
   maxHeight: '90vh',
   maxWidth: 'calc(100vw - 48px)',