+++ packages/client/components/TimelinePriorityTasks.tsx
@@ -1,13 +1,13 @@
 import styled from '@emotion/styled'
-import {Whatshot} from '@mui/icons-material'
-import graphql from 'babel-plugin-relay/macro'
-import {useMemo} from 'react'
 import {
   DragDropContext,
   Droppable,
   type DroppableProvided,
   type DropResult
-} from 'react-beautiful-dnd'
+} from '@hello-pangea/dnd'
+import {Whatshot} from '@mui/icons-material'
+import graphql from 'babel-plugin-relay/macro'
+import {useMemo} from 'react'
 import {useFragment} from 'react-relay'
 import type {TimelinePriorityTasks_viewer$key} from '../__generated__/TimelinePriorityTasks_viewer.graphql'
 import useAtmosphere from '../hooks/useAtmosphere'