+++ packages/client/modules/pages/PageEditor.tsx
@@ -20,7 +20,10 @@ export const PageEditor = (props: Props) => {
   if (!editor) return <div>No editor</div>
   return (
     <>
-      <TipTapEditor editor={editor} className={cn('page-editor flex w-full px-6 delay-300')} />
+      <TipTapEditor
+        editor={editor}
+        className={cn('page-editor relative flex w-full px-6 delay-300')}
+      />
       {isEditable && <StarterActions editor={editor} />}
     </>
   )