+++ packages/client/modules/teamDashboard/components/AgendaList/AgendaList.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled'
+import {DragDropContext, Draggable, Droppable, type DropResult} from '@hello-pangea/dnd'
 import graphql from 'babel-plugin-relay/macro'
 import {useMemo} from 'react'
-import {DragDropContext, Draggable, Droppable, type DropResult} from 'react-beautiful-dnd'
 import {useFragment} from 'react-relay'
 import type {AgendaList_agendaItems$key} from '~/__generated__/AgendaList_agendaItems.graphql'
 import type {AgendaList_meeting$key} from '~/__generated__/AgendaList_meeting.graphql'