+++ packages/server/hocusPocusHandler.ts
@@ -4,7 +4,6 @@ import {redisHocusPocus} from './hocusPocus'
 import './hocusPocus'
 import type {WebSocketBehavior} from 'uWebSockets.js'
 import {HocusPocusWebSocket} from './HocusPocusWebSocket'
-import {Logger} from './utils/Logger'
 import type {SerializedHTTPRequest} from './utils/tiptap/RedisServerAffinity'
 
 export type HocusPocusSocketData = {
@@ -19,9 +18,6 @@ export const hocusPocusHandler: WebSocketBehavior<HocusPocusSocketData> = {
   upgrade(res, req, context) {
     const headers: IncomingHttpHeaders = {}
     req.forEach((key, value) => {
-      if (headers[key]) {
-        Logger.warn(`hocuspocus: overwriting ${key} header. The user may not be able to log in`)
-      }
       headers[key] = value
     })
 
@@ -35,13 +31,6 @@ export const hocusPocusHandler: WebSocketBehavior<HocusPocusSocketData> = {
       url: query ? `${pathname}?${query}` : pathname,
       headers,
       socket: {remoteAddress: ip}
-    }
-    if (!serializedHTTPRequest.headers['cookie']) {
-      const cookie = req.getHeader('cookie')
-      Logger.log('Upgrading yjs websocket without cookie. Should only happen for public pages', {
-        ...serializedHTTPRequest,
-        cookie
-      })
     }
     res.upgrade<HocusPocusSocketData>(
       {